+++ src/LanguageMapOfLists.php
@@ -25,6 +25,16 @@ class LanguageMapOfLists extends LanguageMap
         }
     }
 
+    public function contains($member): bool
+    {
+        foreach ($this->members as $list) {
+            if (in_array($member, $list)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
     /**
      * Return a data structure to serialize this container as JSON.
      */